CoursesApp/requirements.txt
# pip freeze > requirements.txt
# pip install -r requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
mysqlclient==2.2.0
Pillow==10.1.0
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
